Moved request tests to test_internals_request.py
